apps/accounts/serializers.py
"""
Serializers for the accounts app
"""
from rest_framework import serializers
from .models import User, UserProfile


class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ['bio', 'website', 'location', 'birth_date']


class UserSerializer(serializers.ModelSerializer):
    profile = UserProfileSerializer(read_only=True)
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'full_name', 'company_name', 'membership_tier', 'is_verified', 'profile']